Add specific openHPI area
There was a need for a specific area for openHPI.
For courses present in the course catalog data
can be retrieved easily. Only the necessary
attributes edAling and teaches are considered.

The program is theoretically good to go for user
testing. But a database for logging suggestions
would help to track differences to final
assignments.

TODO: Implement some kind of logging of
suggestions.