Fix bug with suggestions
Sometimes the returned raw suggestion von
Gemini did not fit into the template. This
let to exceptions. The main cause of this
behavior (superfluous quotation marks) was
eliminated.
Additionally, try-excepts are set to have the
system running more stable. There will be some
output to trace down the problematic outputs.